requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
docutils==0.20.1
h11==0.14.0
idna==3.6
natsort==8.4.0
outcome==1.3.0.post0
Pygments==2.17.2
PySocks==1.7.1
requests==2.31.0
robotframework==7.0
robotframework-datadriver==1.11.0
robotframework-pabot==2.18.0
robotframework-pythonlibcore==4.3.0
robotframework-requests==0.9.6
robotframework-seleniumlibrary==6.2.0
robotframework-stacktrace==0.4.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
